Ensure dslogic can be stopped.
After starting dslogic, it starts storing samples in memory, and when done it
sends an usb packet with the trigger possition.

This initial fillup can take some time. If user request a session stop in
between, the session was hanging up as not being closed properly.

This commit manages this case and closes the session properly when interrupt
requested by user.

Signed-off-by: Diego F. Asanza <f.asanza@gmail.com>